requirements.txt
attrs==25.3.0
beautifulsoup4==4.12.2
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
exceptiongroup==1.3.0
fake-useragent==2.2.0
Flask==2.3.3
Flask-Cors==4.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.62.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kaggle==1.7.4.5
lxml==4.9.3
MarkupSafe==3.0.2
outcome==1.3.0.post0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
requests==2.31.0
rsa==4.9.1
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
text-unidecode==1.3
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
